backend/requirements.txt
fastapi
uvicorn
requests
deep-translator
python-dotenv
